fix: plugin version restore/delete actions
Swaps the logic for the restore and delete actions in the plugin version history.

The previous implementation incorrectly called the delete endpoint when restoring a version and the restore endpoint when deleting.  This commit corrects this behavior.  It also removes the `withDeleted` flag from the version query as it is no longer needed.