feat: adds read-more directive for long text
Implements a new directive to truncate long text blocks and provide a "read more" button to expand the content.

This improves the user experience by preventing long text descriptions from overflowing their containers and cluttering the UI.  The directive is applied to plugin descriptions and changelogs in the plugin marketplace.